updates
separate into header and implementation
add comments from draft PR as TODO
