--- src/client/dfs/dummy.c
@@ -1,2 +1,0 @@
-/** dummy for compiling */
-#include "daos_fs_sys.h"
